Plan 044 Step 2: extract LRU cache to adapter-cache.ts
Move fetchData cache keying, TTL, LRU eviction, and invalidation into
AdapterCache; the adapter holds an instance with identical semantics.
